update class Topic
modified funciton 'tweet_is_about_topic' by searching the keywords in different order (and add hastags and usernames). This version of change makes no difference for one-word topic but very useful for topic phrases.
However, didn't manage to merge new keywords to the topics for the reason that this function match the topic with only one tweet at a time, which gives difficulty to do so.